Improve netrw usage and appearance
Create useful keymaps to use the preview window easily, create
autocommands to draw icons at the right time, exploiting TextChanged
event but only locally to the netrw buffer so that it does not happen
for every other buffer. Set placement of the cursor avoiding ./ and ../
and close netrw split if the other window opened with it is closed.
Change highlighting groups to differentiate from other buffers.