Lazy load lsp when restoring tmux sessions
Basically, tmux sessions will restore nvim with the option __from_tmux=1
so that the event to lazy load the lsp server is a custom User
TmuxAttach event, which is fired via a tmux hook activating whenever
a session is attached. In this way, lsp servers are not attached
whenever the buffer is read, but only once the tmux session is attached,
which means I am working on it.